refactor : 필터링 조건 함수분리
기존 sort.ts 파일에서 정렬에 대한 함수만을 관리하였던 것을 filterHelper.ts로 변경하여,
모든 필터링 기능을 관리하도록 변경하였습니다.
